Significantly changed the auth.api helper file structure, removing the file and replacing it with a number of more focused files. Likewise the testing for the module is now split into relevant files
