+++ src/pages/ViewBlogs.tsx
@@ -9,7 +9,7 @@ import styles from "./ViewBlogs.module.css"
 
 import { ReactComponent as CreateIcon } from "../assets/images/createIcon.svg"
 import { useNavigate } from "react-router-dom"
-import { hasData } from "../lib/api/auth.api"
+import { hasData } from "../lib/api/helpers/auth/redirectAndClearData"
 
 function getScrollPercentage() {
   const de = document.documentElement
