+++ src/tests/lib/api/helpers/auth/refreshToken.test.ts
@@ -1,16 +1,20 @@
 import { setupServer } from 'msw/node'
 import { rest } from 'msw'
-import { BackendError, BackendResponse } from '../../../lib/commonTypes'
-import { refreshTokens } from '../../../lib/api/auth.api'
+import { BackendError, BackendResponse } from '../../../../../lib/commonTypes'
+import { refreshTokens } from '../../../../../lib/api/helpers/auth/refreshToken'
+import { redirectToSignInAndClearData } from '../../../../../lib/api/helpers/auth/redirectAndClearData'
 
 // Mock the local storage setItem method
 const setItemMock = jest.spyOn(Storage.prototype, "setItem")
 
-// Mock the local storage removeItem method
-const mockRemoveItem = jest.spyOn(Storage.prototype, "removeItem")
-
 // Mock console error
 const consoleErrMock = jest.spyOn(console, "error")
+
+// Mock the redirectToSignInAndClearData function
+jest.mock("../../../../../lib/api/helpers/auth/redirectAndClearData")
+const redirectToSignInAndClearDataMock = jest.mocked(redirectToSignInAndClearData, true).mockImplementation(
+  () => { }
+)
 
 const BASE_PATH = process.env.REACT_APP_BACKEND_SERVER_ADDR
 const USER_ID = "someUserId"
@@ -51,21 +55,10 @@ describe("refreshTokens", () => {
   })
 
   describe("When the client does not have a valid refresh token", () => {
-    let locationSpy: jest.SpyInstance<Location, []>
-    const setHrefMock = jest.fn((value: string) => { })
-
     // Run setups
     beforeEach(async () => {
-      // Mock the window.location object
-      locationSpy = jest.spyOn(window, "location", "get").mockImplementation(() => {
-        return {
-          set href(value: string) { setHrefMock(value) }
-        } as Location
-      })
-
       // Reset mocks
       consoleErrMock.mockReset()
-      mockRemoveItem.mockReset()
 
       // Setup route
       server.use(
@@ -81,21 +74,12 @@ describe("refreshTokens", () => {
       await refreshTokens()
     })
 
-    afterAll(() => {
-      // Restore the window.location mock
-      locationSpy.mockRestore()
-    })
-
     it("Outputs an error to console error", () => {
       expect(consoleErrMock).toHaveBeenCalledWith("Error: Refresh token is not valid")
     })
 
-    it("Clears client data", () => {
-      expect(mockRemoveItem).toHaveBeenCalledWith("userId")
-    })
-
-    it("Redirects the client to the sign in page", () => {
-      expect(setHrefMock).toHaveBeenCalledWith(process.env.REACT_APP_FRONTEND_SERVER_ADDR + "signin")
+    it("Clears client data and redirect to the sign in page", () => {
+      expect(redirectToSignInAndClearDataMock).toHaveBeenCalledTimes(1)
     })
   })
 })
